Revert "adicionando autenticação por jwt"
