--- src/main/java/br/com/rotaract/secretaria/config/SecurityConfigurations.java
@@ -1,57 +1,0 @@
-package br.com.rotaract.secretaria.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import br.com.rotaract.secretaria.filter.AutenticacaoTokenFilter;
-import br.com.rotaract.secretaria.repository.AssociadoRepository;
-import br.com.rotaract.secretaria.service.AutenticacaoService;
-import br.com.rotaract.secretaria.service.TokenService;
-
-@EnableWebSecurity
-@Configuration
-public class SecurityConfigurations extends WebSecurityConfigurerAdapter{
-
-	@Autowired
-	private AutenticacaoService autenticacaoService;
-	
-	@Autowired
-	private TokenService tokenService;
-	
-	@Autowired
-	private AssociadoRepository associadoRepository;
-	
-	@Override
-	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.userDetailsService(autenticacaoService).passwordEncoder(new BCryptPasswordEncoder());
-	}
-	
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		http.authorizeRequests()
-			.antMatchers("/auth").permitAll()
-			//Criar usuario admin e remover essa linha
-			.antMatchers(HttpMethod.POST, "/associado").permitAll()
-			.antMatchers("/").permitAll()
-			.anyRequest().authenticated()
-			.and().cors().and().csrf().disable()
-			.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-			.and().addFilterBefore(new AutenticacaoTokenFilter(tokenService, associadoRepository), UsernamePasswordAuthenticationFilter.class);
-	}
-
-	@Override
-	@Bean
-	protected AuthenticationManager authenticationManager() throws Exception {
-		return super.authenticationManager();
-	}
-}
