--- src/main/java/br/com/rotaract/secretaria/dto/TokenDto.java
@@ -1,29 +1,0 @@
-package br.com.rotaract.secretaria.dto;
-
-public class TokenDto {
-
-	private String token;
-	private String tipo;
-
-	public TokenDto(String token, String tipo) {
-		this.token = token;
-		this.tipo = tipo;
-	}
-
-	public String getToken() {
-		return token;
-	}
-
-	public void setToken(String token) {
-		this.token = token;
-	}
-
-	public String getTipo() {
-		return tipo;
-	}
-
-	public void setTipo(String tipo) {
-		this.tipo = tipo;
-	}
-
-}
