+++ src/main/java/br/com/rotaract/secretaria/model/Associado.java
@@ -1,8 +1,6 @@
 package br.com.rotaract.secretaria.model;
 
 import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.List;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -10,9 +8,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import br.com.rotaract.secretaria.constant.StatusAssociado;
 import lombok.AllArgsConstructor;
@@ -22,9 +17,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Table(name = "associado")
 @Entity
-public class Associado implements UserDetails{
-
-	private static final long serialVersionUID = 1L;
+public class Associado {
 
 	@Id
 	@Column(name = "id_associado")
@@ -39,12 +32,6 @@ public class Associado implements UserDetails{
 	@Column(name = "padrinho")
 	private String padrinho;
 
-	@Column(name = "email")
-	private String email;
-
-	@Column(name = "senha")
-	private String senha;
-
 	@OneToOne
 	@JoinColumn(name = "id_pessoa", nullable = false)
 	private Pessoa pessoa;
@@ -52,41 +39,6 @@ public class Associado implements UserDetails{
 	@OneToOne
 	@JoinColumn(name = "id_cargo", nullable = false)
 	private Cargo cargo;
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return List.of(this.cargo);
-	}
-
-	@Override
-	public String getPassword() {
-		return this.senha;
-	}
-
-	@Override
-	public String getUsername() {
-		return this.email;
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
 
 	public Long getRI() {
 		return RI;
@@ -100,8 +52,8 @@ public class Associado implements UserDetails{
 		return status;
 	}
 
-	public void setStatus(StatusAssociado status) {
-		this.status = status;
+	public void setStatus(StatusAssociado ativo) {
+		this.status = ativo;
 	}
 
 	public LocalDateTime getDataAdmissao() {
@@ -118,22 +70,6 @@ public class Associado implements UserDetails{
 
 	public void setPadrinho(String padrinho) {
 		this.padrinho = padrinho;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getSenha() {
-		return senha;
-	}
-
-	public void setSenha(String senha) {
-		this.senha = senha;
 	}
 
 	public Pessoa getPessoa() {
