+++ src/main/java/br/com/rotaract/secretaria/model/Pessoa.java
@@ -42,6 +42,9 @@ public class Pessoa {
 	@Column(name = "restricoes_medicas")
 	private String restricoesMedicas;
 
+	@Column(name = "email")
+	private String email;
+
 	@Column(name = "telefone")
 	private String telefone;
 
@@ -50,6 +53,9 @@ public class Pessoa {
 
 	@Column(name = "nome_cracha")
 	private String cracha;
+
+	@Column(name = "senha")
+	private String senha;
 
 	@OneToOne
 	@JoinColumn(name = "id_endereco", nullable = false)
@@ -107,6 +113,14 @@ public class Pessoa {
 		this.restricoesMedicas = restricoesMedicas;
 	}
 
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
 	public String getTelefone() {
 		return telefone;
 	}
@@ -129,6 +143,14 @@ public class Pessoa {
 
 	public void setCracha(String cracha) {
 		this.cracha = cracha;
+	}
+
+	public String getSenha() {
+		return senha;
+	}
+
+	public void setSenha(String senha) {
+		this.senha = senha;
 	}
 
 	public Endereco getEndereco() {
