+++ src/main/java/br/com/rotaract/secretaria/repository/AssociadoRepository.java
@@ -1,7 +1,5 @@
 package br.com.rotaract.secretaria.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,7 +7,5 @@ import br.com.rotaract.secretaria.model.Associado;
 
 @Repository
 public interface AssociadoRepository extends JpaRepository<Associado, Long>{
-	
-	Optional<Associado> findByEmail(String email);
 
 }
