--- src/main/java/br/com/rotaract/secretaria/service/TokenService.java
@@ -1,51 +1,0 @@
-package br.com.rotaract.secretaria.service;
-
-import java.util.Date;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Service;
-
-import br.com.rotaract.secretaria.model.Associado;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-
-@Service
-public class TokenService {
-	
-	@Value("${autenticacao.jwt.expiration}")
-	private String expiration;
-	
-	@Value("${autenticacao.jwt.secret}")
-	private String secret;
-
-	public String gerarToken(Authentication authentication) {
-		
-		Associado logado = (Associado) authentication.getPrincipal();
-		Date hoje = new Date();
-		Date expiracao = new Date(hoje.getTime() + Long.parseLong(expiration));
-		
-		return Jwts.builder()
-				.setIssuer("")
-				.setSubject(logado.getRI().toString())
-				.setIssuedAt(hoje)
-				.setExpiration(expiracao)
-				.signWith(SignatureAlgorithm.HS256, secret)
-				.compact();
-	}
-
-	public boolean isTokenValido(String token) {
-		try {
-			Jwts.parser().setSigningKey(this.secret).parseClaimsJws(token);
-			return true;
-		} catch (Exception e) {
-			return false;
-		}
-	}
-
-	public Long getIdAssociado(String token) {
-		Claims body = Jwts.parser().setSigningKey(this.secret).parseClaimsJws(token).getBody();
-		return Long.parseLong(body.getSubject());
-	}
-}
