+++ src/main/resources/db/changelog/changes/06-create-pessoa.xml
@@ -16,9 +16,11 @@
 			OCUPACAO VARCHAR(25),
 			DATA_NASCIMENTO DATE,
 			RESTRICOES_MEDICAS VARCHAR(50),
+		    EMAIL VARCHAR(50),
 		    TELEFONE VARCHAR(11),
 		    CPF VARCHAR(11),
 		    NOME_CRACHA VARCHAR(50),
+		    SENHA VARCHAR(25),
 			PRIMARY KEY (ID_PESSOA),
 			FOREIGN KEY (ID_ENDERECO) REFERENCES ENDERECO(ID_ENDERECO) ON UPDATE CASCADE ON DELETE CASCADE
 		);
