+++ src/main/resources/db/changelog/changes/10-create-associado.xml
@@ -16,9 +16,7 @@
 			ID_PESSOA BIGINT,
 			STATUS_ASSOCIADO VARCHAR(25),
 			DATA_ADMISSAO DATE,
-			PADRINHO VARCHAR(50),
-			EMAIL VARCHAR(50),
-			SENHA VARCHAR(100),
+			PADRINHO VARCHAR(50),	
 			PRIMARY KEY (ID_ASSOCIADO),
 			FOREIGN KEY (ID_CARGO) REFERENCES CARGO(ID_CARGO) ON UPDATE CASCADE ON DELETE CASCADE,
 			FOREIGN KEY (ID_PESSOA) REFERENCES PESSOA(ID_PESSOA) ON UPDATE CASCADE ON DELETE CASCADE,
